refactor: move queryservice and rpcclient to go/common
Previously these were under go/multipooler, but they're shared code that's used by other services (multiorch, multigateway, etc.). Keeping them under go/common will make that more clear
