+++ go/common/rpcclient/fake_client.go
