+++ go/multigateway/poolergateway/grpc_query_service.go
@@ -20,7 +20,7 @@ import (
 	"io"
 	"log/slog"
 
-	"github.com/multigres/multigres/go/multipooler/queryservice"
+	"github.com/multigres/multigres/go/common/queryservice"
 	"github.com/multigres/multigres/go/pb/multipoolerservice"
 	"github.com/multigres/multigres/go/pb/query"
